fix clang warnings

+++ include/core_bits.h
@@ -19,7 +19,7 @@ constexpr inline To bitCast(const From& src) noexcept {
 
 // Least Significant N Bits are equal to bitSeq
 constexpr bool leastSignificantNBits(u8 v, u8 bitSeq, u8 n) {
-    u8 mask = ~u8(core::limitMax<u8>() << n);
+    u8 mask = u8(~u8(core::limitMax<u8>() << n));
     v = (v & mask);
     bool ret = (v == bitSeq);
     return ret;
